refactor/bandaid: split `FirstAidKit` from `Bandaid` file
